dbg: Break up debugable into modules
