+++ core/src/context.rs
@@ -354,7 +354,7 @@ impl<'a, 'gc> UpdateContext<'a, 'gc> {
 
     pub fn avm_trace(&self, message: &str) {
         let dbg = &self.debugger;
-        dbg.submit_debug_message(crate::debugable::DebugMessageOut::LogTrace(
+        dbg.submit_debug_message(crate::debug::debug_message_out::DebugMessageOut::LogTrace(
             message.to_string(),
         ));
         self.log.avm_trace(&message.replace('\r', "\n"));
