+++ core/src/display_object.rs
@@ -1427,19 +1427,13 @@ pub trait TDisplayObject<'gc>:
             classname = format!("{:?}", o.base().debug_class_name());
         }
 
-        let path = self.slash_path().to_string();
-        let v = self.visible();
-
         println!(
-            "{} rel({},{}) abs({},{}) depth({}), path({}) vis({}) {} {} {}",
+            "{} rel({},{}) abs({},{}) {} {} {}",
             " ".repeat(depth),
             self.x(),
             self.y(),
             bounds.x_min.to_pixels(),
             bounds.y_min.to_pixels(),
-            self.depth(),
-            path,
-            v,
             classname,
             self.name(),
             self_str
@@ -1508,7 +1502,7 @@ pub trait TDisplayObject<'gc>:
         None
     }
 
-    fn as_debuggable(self) -> Option<crate::debugable::Debuggable<'gc>> {
+    fn as_debuggable(self) -> Option<crate::debug::debuggable::Debuggable<'gc>> {
         None
     }
