+++ desktop/Cargo.toml
@@ -28,11 +28,10 @@ anyhow = "1.0"
 bytemuck = "1.13.0"
 os_info = { version = "3", default-features = false }
 tungstenite = { version = "0.18.0"}
+serde_json = { version = "1.0" }
 
 # Deliberately held back to match tracy client used by profiling crate
 tracing-tracy = { version = "=0.10.0", optional = true }
-serde = { version = "1.0.150", features = ["derive"] }
-serde_json = { version = "1.0", features = ["preserve_order"] }
 
 [target.'cfg(windows)'.dependencies]
 winapi = "0.3.9"
